gcode
Module buffers

Buffer Management.

This module is mainly intended for use cases when the amount of space that can be consumed by buffers needs to be defined at compile time. For most users, the DefaultBuffers alias should be suitable.

For most end users it is probably simpler to determine a “good enough” buffer size and create type aliases of GCode and Line for that size.

Structs

CapacityError

An error returned when Buffer::try_push() fails.

Enums

SmallFixedBuffers

The smallest usable set of Buffers.

VecBuffers

A Buffers implementation which uses std::vec::Vec for storing items.

Traits

Buffer

Something which can store items sequentially in memory. This doesn’t necessarily require dynamic memory allocation.

Buffers

A set of type aliases defining the types to use when storing data.

Type Aliases

DefaultArguments

The default Buffer to use for a GCode’s arguments.

DefaultBuffers

The default buffer type for this platform.