gcode::core
Type Alias ControlFlow

pub type ControlFlow<T> = ControlFlow<(), T>;

Return type for visitor methods that may either continue with a child visitor or pause parsing. See the module-level docs for the control-flow model.

Aliased Type

enum ControlFlow<T> {
    Continue(T),
    Break(()),
}

Variants

Continue(T)

Move on to the next phase of the operation as normal.

Break(())

Exit the operation without running subsequent phases.