gcode
Function parse

pub fn parse<'input>(src: &'input str) -> impl Iterator<Item = GCode> + 'input

Parse each GCode in some text, ignoring any errors that may occur or Comments that are found.

This function is probably what you are looking for if you just want to read the GCode commands in a program. If more detailed information is needed, have a look at full_parse_with_callbacks().