Enum gcode::high_level::GCode

pub enum GCode {
    G00 {
        to: Point,
        feed_rate: Option<f32>,
    },
    G01 {
        to: Point,
        feed_rate: Option<f32>,
    },
    G02 {
        to: Point,
        feed_rate: Option<f32>,
        radius: Option<f32>,
        centre: Option<Point>,
    },
    G04 {
        seconds: f32,
    },
}

Variants
G00

Rapid Linear Motion

Fields of G00

to: Point

feed_rate: Option<f32>

G01

Linear Motion at Feed Rate

Fields of G01

to: Point

feed_rate: Option<f32>

G02

Clockwise Arc at Feed Rate

Note

positive radius indicates that the arc turns through 180 degrees or less, while a negative radius indicates a turn of 180 degrees to 359.999 degrees.

Fields of G02

to: Point

feed_rate: Option<f32>

radius: Option<f32>

centre: Option<Point>

G04

Dwell - wait for a number of seconds

Fields of G04

seconds: f32