Function gcode::high_level::type_check

pub fn type_check(line: Line) -> Result<Line>

Convert the loosely typed low_level::Line into its more strongly typed representation.

Note that as a result of this process you tend to lose line information. It's assumed that if you get this far in the pipeline you've already dealt with errors.