Module gcode::parser

The main parsing module for this crate.

Structs

Args
A good ol' bag-o-floats which contains all the possible arguments and their values.

Command
A single command.

Parser
A parser which turns a stream of gcode tokens into a stream of commands.

Enums

CommandKind
The type of command.

Line
A single line of gcode.

Number
A type which can either be an integer or a float.