Struct gcode::Block

pub struct Block<'input> { /* fields omitted */ }

A block containing Gcode commands and/or comments.

Methods

impl<'input> Block<'input>

pub fn is_empty(&self) -> bool

Is this block empty?

pub fn into_commands(self) -> impl Iterator<Item = Gcode>

Convert this Block into a stream of Gcode commands.

pub fn commands(&self) -> &[Gcode]

The Gcodes in this block.

pub fn commands_mut(&mut self) -> &mut [Gcode]

Get a mutable reference to the Gcodes in this block.

pub fn comments(&self) -> &[Comment<'input>]

The comments in this block.

pub fn span(&self) -> Span

The block's location.

pub fn line_number(&self) -> Option<usize>

The block's line number (e.g. N42), if provided.

pub fn with_line_number(&mut self, number: usize, span: Span) -> &mut Self

Set the Block's line number.

pub fn deleted(&self) -> bool

Does this block start with a block delete?

pub fn delete(&mut self, delete: bool)

Set the block delete flag.

pub fn push_comment(&mut self, comment: Comment<'input>)

Add a Comment to the Block.

pub fn push_command(&mut self, command: Gcode)

Add a Gcode to the Block.