Struct gcode::Gcode

pub struct Gcode { /* fields omitted */ }

A single gcode command.

Methods

impl Gcode

pub const MAX_ARGUMENT_COUNT: usize

The maximum number of arguments in a single [Gcode] command when compiled without the std feature.

Each [Gcode] can contain 16 [Argument]s.

pub fn new(mnemonic: Mnemonic, number: f32) -> Gcode

Create a new Gcode.

Panics

A negative number is invalid.

pub fn mnemonic(&self) -> Mnemonic

The general command category.

pub fn major_number(&self) -> usize

Get the command's major number.

pub fn minor_number(&self) -> Option<usize>

Get the command's minor number, of there is one.

pub fn args(&self) -> &[Argument]

Access any arguments attached to this Gcode.

pub fn args_mut(&mut self) -> &mut [Argument]

Mutably access the arguments attached to this Gcode.

pub fn with_line_number(self, number: usize) -> Self

Set the line number.

pub fn with_minor_number(self, number: usize) -> Self

Set the minor number.

pub fn with_argument(self, arg: Argument) -> Self

Add an argument to the [Gcode], removing any previous arguments with the same letter.

pub fn with_span(self, span: Span) -> Self

Set the Gcode's Span.

pub fn push_argument(&mut self, arg: Argument)

Add an argument to the Gcode.

pub fn remove_argument(&mut self, letter: char) -> Option<Argument>

Remove the first argument with the specified letter.

pub fn span(&self) -> Span

Get the Gcode's location.

pub fn line_number(&self) -> Option<usize>

Get the Gcode's line number, if one was provided.

pub fn value_for(&self, letter: char) -> Option<f32>

Get the value for a particular argument letter, if there is one.