Trait gcode::transforms::Predicate

pub trait Predicate<T: ?Sized> {
    fn evaluate(&mut self, item: &T) -> bool;

    fn and<P>(self, other: P) -> And<Self, P>
    where
        P: Predicate<T>,
        Self: Sized,
    { ... }

    fn or<P>(self, other: P) -> Or<Self, P>
    where
        P: Predicate<T>,
        Self: Sized,
    { ... }
}

A generic predicate.

The intention behind this helper trait is to make applying transformations to a subset of inputs more ergonomic.

use gcode::{Argument, Gcode, Mnemonic, Span};
use gcode::transforms::Predicate;

let g90 = Gcode::new(Mnemonic::General, 90.0);
let g00 = Gcode::new(Mnemonic::General, 0.0)
    .with_argument(Argument::new('X', 500.0, Span::placeholder()));
let m6 = Gcode::new(Mnemonic::Miscellaneous, 6.0);
let m0 = Gcode::new(Mnemonic::Miscellaneous, 0.0);

// mnemonics can be used to match any gcode with that mnemonic
let mut matches_g = Mnemonic::General;
assert!(matches_g.evaluate(&g90));
assert!(!matches_g.evaluate(&m6));

// You can also match on a major number
let mut matches_0 = 0;
assert!(matches_0.evaluate(&g00));
assert!(matches_0.evaluate(&m0));
assert!(!matches_0.evaluate(&g90));

// predicates can also be combined to make more specific predicates
let mut matches_g90 = Mnemonic::General.and(90);
assert!(matches_g90.evaluate(&g90));
assert!(!matches_g90.evaluate(&m0));
assert!(!matches_g90.evaluate(&g00));

// you can also combine mutation and closures in interesting ways
let mut count = 0;
let mut matches_every_other_call = |_: &Gcode| {
    count += 1;
    count % 2 == 0
};
assert!(!matches_every_other_call.evaluate(&g00));
assert!(matches_every_other_call.evaluate(&g90));
assert!(!matches_every_other_call.evaluate(&g90));
assert!(matches_every_other_call.evaluate(&g00));

// What about only matching on a G-code with an `X` parameter?
let mut x = 'X';
assert!(x.evaluate(&g00));

Required Methods

fn evaluate(&mut self, item: &T) -> bool

Evaluate the predicate based on the provided input.

Provided Methods

fn and<P>(self, other: P) -> And<Self, P> where
P: Predicate<T>,
Self: Sized,

Create a new [Predicate<T>] which is effectively this.evaluate(item) && other.evaluate(item).

fn or<P>(self, other: P) -> Or<Self, P> where
P: Predicate<T>,
Self: Sized,

Create a new [Predicate<T>] which is effectively this.evaluate(item) || other.evaluate(item).