Crate gcode

A crate for parsing gcodes without relying on std.

The crate uses a pipeline pattern and iterators to implement a zero-allocation lexer and parser.

Pipeline Stages

The data goes through a series of transformations before it arrives at its final strongly-typed representation. Each transform uses the Iterator trait to lazily evaluate commands, so if you aren't reading from local memory (i.e. an attached SD card), if you're reading data from a serial port or other network device you'll probably want to wrap it in some sort of buffered reader.

Stage 	Transform 	Transformer 	Description

1 	char -> Token 	Tokenizer 	Lexical analysis (tokenizing)

2 	Token -> low_level::Line 	BasicParser 	Initial parsing

3 	Line -> high_level::Line 	type_check() 	Type-checking and strong typing

Examples

You can manually exercise the entire pipeline as follows:

use gcode::{Tokenizer, BasicParser, type_check};

let src = "G00 X10.0 Y20.0; G00 Z-10.0; G01 X55.2 Y-32.0 F500;";

// Construct a tokenizer and pass it our source code.
let lexer = Tokenizer::new(src.chars());

// Ignore any errors we encounter to get a stream of `Token`s.
let tokens = lexer.filter_map(|t| t.ok());

// Construct a parser which takes in the tokens.
let parser = BasicParser::new(tokens);

// Skip all parsing errors and then apply type checking, skipping errors again
let lines = parser.filter_map(|l| l.ok())
                  .map(|l| type_check(l))
                  .filter_map(|l| l.ok());

for line in lines {
    println!("{:?}", line);
}

Alternatively, using the nightly feature flag, you can just use the parse() function.

use gcode::parse;

let src = "G00 X10.0 Y20.0; G00 Z-10.0; G01 X55.2 Y-32.0 F500;";

for line in parse(src.chars()) {
    println!("{:?}", line);
}

Feature Flags

At the moment only one feature flag is available.

nightly enables the parse function. This is a convenience function which will convert a stream of chars into a stream of high_level::Lines. It relies on conservative_impl_trait.

Reexports

pub use lexer::Tokenizer;

pub use lexer::Span;

pub use low_level::BasicParser;

pub use high_level::type_check;

Modules

high_level
The high level interpreting of parsed Commands as their particular G and M codes and applying strong typing to their arguments.

lexer
A module for turning raw gcode into tokens to be processed by the parser.

low_level
The first stage of parsing which turns tokens into a basic gcode representation.

Enums

Error
The error type.

Type Definitions

Result
An alias for the Result type.