Crate gcode

A crate for parsing gcodes without relying on std.

The crate uses iterators extensively to implement a zero-allocation lexer and parser.

Examples

You can manually exercise the entire pipeline as follows:

use gcode::{Tokenizer, Parser};

let src = "G00 X10.0 Y20.0; G00 Z-10.0; G01 X55.2 Y-32.0 F500;";

// Construct a tokenizer and pass it our source code.
let lexer = Tokenizer::new(src.chars());

// Ignore any errors we encounter to get a stream of `Token`s.
let tokens = lexer.filter_map(|t| t.ok());

// Construct a parser which takes in the tokens.
let parser = Parser::new(tokens);

// Skip all parsing errors and then apply type checking, skipping errors again
let lines = parser.filter_map(|l| l.ok());

for line in lines {
    println!("{:?}", line);
}

Reexports

pub use parser::Parser;

pub use lexer::Tokenizer;

pub use lexer::Span;

pub use low_level::BasicParser;

Modules

lexer
A module for turning raw gcode into tokens to be processed by the parser.

low_level 	 [
Deprecated
]
The first stage of parsing which turns tokens into a basic gcode representation.

parser
The main parsing module for this crate.

Enums

Error
The error type.

Type Definitions

Result
An alias for the Result type.