Crate gcode

A gcode parsing library designed for no_std environments.

Examples

The gcode API is extremely minimal, containing a single parse() function which takes a string and returns an iterator over the Gcodes it contains.

Parsing is done on a best-effort basis, with any syntax errors being silently ignored. Because of this, you'll probably want to ensure a file is well formed.

Note: The examples/ folder in this project's repository contains a simple program that will read an input file and print out any Gcodes it sees.

use gcode::Mnemonic;

let src = "O1000
    T1 M6
    G90 
    G01 X-75 Y-75 S500 M3 
    G43 Z100 H1
    G01 Z5
    N20 G01 Z-20 F100";

let mut lines = gcode::parse(src);

let program_number = lines.next().unwrap();
assert_eq!(program_number.major_number(), 1000);

let tool_change = lines.next().unwrap();
assert_eq!(tool_change.mnemonic(), Mnemonic::ToolChange);
assert_eq!(tool_change.major_number(), 1);

// skip the M6 and G90
let _ = lines.next();
let _ = lines.next();

let g01 = lines.next().unwrap();
assert_eq!(g01.major_number(), 1);
assert_eq!(g01.args().len(), 3);
assert_eq!(g01.value_for('X'), Some(-75.0));

let rest: Vec<_> = lines.collect();
assert_eq!(rest.len(), 4);
assert_eq!(rest[3].line_number(), Some(20));

Modules

ffi
A FFI interface to the gcode library.

Structs

Gcode
A single command in the gcode programming language.

Parser
A gcode parser which is extremely permissive in what input it will accept.

Span
A set of byte indices which correspond to the location of a substring in a larger piece of text.

Word
A single Word in the gcode language (e.g. X-12.3).

Enums

Mnemonic
A general command category.

Constants

MAX_ARGS
The maximum number of arguments a Gcode can have.

Functions

parse
Parse a string of text into a stream of Gcodes.